Added Powershell PromptingFunctions module
Added the PromptingFunctions.psm1 module. This is a Powershell module
for defining a hierarchy of steps (to be executed manually or
programatically), where the user is asked whether the next step should
be executed or not. This was used to standardize a checklist of steps
for planning the Fantasy League team each week, and capture the results
after each weekend.